python-service/requirements.txt
praat-parselmouth==0.4.6
numpy>=1.26.0
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.18
requests==2.31.0
